Add home page image with Material UI components.
